Add files via upload
For This task, 
At first I create a simple login screen with two TextField widgets for 'username' and 'password' using TextEditingController and the ElevatedButton that is 'Login button'.
Added _correctUsername and _correctPassword variables for login validation. If the credentials don’t match, an error SnackBar displays "Invalid username or password."
After the loggin is successful, there opens a screen named "Home Screen". For Navigate to HomeScreen there i use the _handleLogin function, when the login credentials match the hardcoded username and password, we use Navigator.push to navigate to HomeScreen.